Extract ClassIcon and StoreIcons components
