+++ src/app/character-tile/StoreIcons.m.scss
@@ -1,0 +1,5 @@
+.classIcon {
+  width: 16px;
+  height: 16px;
+  margin-right: 4px;
+}
