+++ src/app/item-popup/DesktopItemActions.m.scss
@@ -26,13 +26,6 @@
   :global(.app-icon) {
     color: #eee;
   }
-
-  .classIcon {
-    filter: invert(1);
-    width: 16px;
-    height: 16px;
-    margin-right: 4px;
-  }
 }
 
 .itemTagSelector {
@@ -68,9 +61,6 @@
     color: black;
     :global(.app-icon) {
       color: black;
-    }
-    .classIcon {
-      filter: none;
     }
   }
 }
