+++ src/app/organizer/ItemTable.m.scss
@@ -249,6 +249,8 @@ $toolbarHeight: 41px - 16px + 8px;
   white-space: nowrap;
   img {
     margin-right: 4px;
+    height: 16px;
+    width: 16px;
   }
 }
